refactor(lib): migrate data imports to domain modules
Route every callsite through the new post/tag/page data modules
instead of the consolidated data/posts entry point.
