+++ src/app/(ArchivePage)/archive/page.tsx
@@ -7,8 +7,8 @@ import Heading from '@/components/UI/Heading';
 import { Stack } from '@/components/UI/Layout';
 import { Title } from '@/components/UI/Title';
 import { SITE_URL } from '@/constants';
-import { getPostsListJson } from '@/lib/data/posts';
 import { getCollectionPageStructured } from '@/lib/domain/json-ld';
+import { getPostsListJson } from '@/lib/post/data';
 import { getData } from './_lib';
 
 const posts = getPostsListJson();
