+++ src/app/(ArchivePage)/tags/[slug]/page.tsx
@@ -11,8 +11,8 @@ import { Spinner } from '@/components/UI/Spinner';
 import { Title } from '@/components/UI/Title';
 import { SITE_NAME, SITE_URL, TAG_VIEW_LIMIT } from '@/constants';
 import { isProduction } from '@/lib/config/environment';
-import { getTagsWithCount } from '@/lib/data/posts';
 import { getCollectionPageStructured } from '@/lib/domain/json-ld';
+import { getTagsWithCount } from '@/lib/tag/data';
 
 type Params = Promise<{ slug: string }>;
