+++ src/app/(PostPage)/[slug]/page.tsx
@@ -16,8 +16,8 @@ import { Stack } from '@/components/UI/Layout';
 import { Container } from '@/components/UI/Layout/Container';
 import { AUTHOR_NAME } from '@/constants';
 import { buildId } from '@/lib/config/environment';
-import { getPostsListJson } from '@/lib/data/posts';
 import { getBlogPostingStructured, getBreadcrumbStructured, getDescriptionText } from '@/lib/domain/json-ld';
+import { getPostsListJson } from '@/lib/post/data';
 import { getOgpImage, getPermalink } from '@/lib/utils/url';
 import { getPostPageData } from './lib/services';
