+++ src/app/_lib/getPopularPost.ts
@@ -1,4 +1,4 @@
-import { type getPostsListJson, getPostsPopular } from '@/lib/data/posts';
+import { type getPostsListJson, getPostsPopular } from '@/lib/post/data';
 import type { PostSummary } from '@/types/source';
 import { IGNORE_SLUGS } from './constants';
 import { getDateAndUpdatedToSimpleFormat } from './getDateAndUpdatedToSimpleFormat';
